merge

+++ IMS-Starter/src/main/java/com/qa/ims/controller/CustomerController.java
@@ -9,10 +9,10 @@ import com.qa.ims.persistence.dao.CustomerDAO;
 import com.qa.ims.persistence.domain.Customer;
 import com.qa.ims.utils.Utils;
 
-/**
- * Takes in customer details for CRUD functionality
- *
- */
+
+ //Takes in customer details for CRUD functionality
+
+
 public class CustomerController implements CrudController<Customer> {
 
 	public static final Logger LOGGER = LogManager.getLogger();
@@ -26,9 +26,9 @@ public class CustomerController implements CrudController<Customer> {
 		this.utils = utils;
 	}
 
-	/**
-	 * Reads all customers to the logger
-	 */
+
+	 //Reads all customers to the logger
+
 	@Override
 	public List<Customer> readAll() {
 		List<Customer> customers = customerDAO.readAll();
@@ -38,45 +38,92 @@ public class CustomerController implements CrudController<Customer> {
 		return customers;
 	}
 
-	/**
-	 * Creates a customer by taking in user input
-	 */
+
+	 //Creates a customer by taking in user input
+
 	@Override
 	public Customer create() {
-		LOGGER.info("Please enter a first name");
-		String firstName = utils.getString();
-		LOGGER.info("Please enter a surname");
-		String surname = utils.getString();
-		Customer customer = customerDAO.create(new Customer(firstName, surname));
-		LOGGER.info("Customer created");
+		Customer customer = null;
+
+		while (customer == null) {
+
+			LOGGER.info("Please enter a first name\n======================= \nSelect '=' to Cancel ");
+			String firstName = utils.getString();
+			if (firstName.equals("=")) {
+				LOGGER.info("REQUEST CANCELLED");
+				break;
+			}
+			LOGGER.info("Please enter a surname \n Select '=' to Return To Cancel Request \nSelect '-' To Cancel Request");
+			String surname = utils.getString();
+			if (surname.equals("=")) {
+				LOGGER.info("Returned To Previous Step\n==============================");
+				continue;
+			}
+			if(surname.equals("-")) {
+				LOGGER.info("REQUEST CANCELLED");
+				break;
+			}
+			customer = customerDAO.create(new Customer(firstName, surname));
+			LOGGER.info("Customer created");
+			return customer;
+		}
 		return customer;
 	}
 
-	/**
-	 * Updates an existing customer by taking in user input
-	 */
+
+	 //Updates an existing customer by taking in user input
+
 	@Override
 	public Customer update() {
-		LOGGER.info("Please enter the id of the customer you would like to update");
-		Long id = utils.getLong();
-		LOGGER.info("Please enter a first name");
-		String firstName = utils.getString();
-		LOGGER.info("Please enter a surname");
-		String surname = utils.getString();
-		Customer customer = customerDAO.update(new Customer(id, firstName, surname));
-		LOGGER.info("Customer Updated");
+
+		Customer customer = null;
+		Customer firstNames = null;
+
+		while (customer == null) {
+
+			LOGGER.info("Please enter the ID of the customer you would like to update");
+			Long id = utils.getLong();
+			if (id == 0) {
+				LOGGER.info("REQUEST CANCELLED");
+				return null;
+			}
+
+				while (firstNames == null) {
+
+				LOGGER.info("Please enter a first name \nSelect R to return to previous step");
+				String firstName = utils.getString();
+				if (firstName.equals("r")) {
+					continue;
+				}
+
+				LOGGER.info("Please enter a surname");
+				String surname = utils.getString();
+				if (surname.equals("r")) {
+					continue;
+				}
+				customer = customerDAO.update(new Customer(id, firstName, surname, surname, 0));
+				LOGGER.info("Customer Updated");
+				return null;
+			}
+
+		}
+
 		return customer;
 	}
 
-	/**
-	 * Deletes an existing customer by the id of the customer
-	 * 
-	 * @return
-	 */
+
+	 //Deletes an existing customer by the id of the customer
+
+	 //@return
+
 	@Override
 	public int delete() {
 		LOGGER.info("Please enter the id of the customer you would like to delete");
 		Long id = utils.getLong();
+		if (id == 0) {
+			LOGGER.info("REQUEST CANCELLED");
+			return 0;
+		}
 		return customerDAO.delete(id);
 	}
